Spanish translation of troubleshooting.md file
Observation: You should check if this paragraph is correct

>They must be 644. If any of those file has permission 664 you need to change it to 644. Use your hosting's control panel or chmod command.

I changed 'file has' for 'files have' because of the 'those' before. But I think that you meant ''If any of those files have NOT permission 664 you need to change it to 644''. I didn´t do any modifications yet, ignore this observation if I was mistaken.